% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt_branching.r
\name{branchcut}
\alias{branchcut}
\title{Split a branch into three}
\usage{
branchcut(dpt, bid, b)
}
\arguments{
\item{dpt}{Pseudo time distances for each of the three branches     (\eqn{n \times 3} matrix of double)}

\item{bid}{Cell indices in ascending pseudotime distance per branch (\eqn{n \times 3} matrix of integer)}

\item{b}{Index of branch to cut (1, 2, or 3)}
}
\description{
Assign cells to a branch b by maximizing finite Kendall correlation of dpts of the
other two branches on b + finite Kendall anticorrelation of dpts on the other branches.
}
\details{
Both dpt and indb contain an entry for each cell, as this function performs the cut.
}

