% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/guo-data.r
\docType{data}
\name{data-guo}
\alias{data(guo)}
\alias{data-guo}
\alias{guo}
\title{Guo at al. mouse embryonic stem cell qPCR data}
\format{A data frame with 492 rows and 49 variables, the first 48 of which are Ct values for gene expression data}
\usage{
data(guo)
}
\value{
a \eqn{429 \times 49} \code{data.frame} with the first 48 columns being qPCR Ct values and the last one the "divisions" annotation.
}
\description{
Gene expression data of 48 genes and an annotation column \code{$division} containing the number of divisions after which the cells were harvested.
}
\details{
The data is normalized using the mean of two housekeeping genes and the LoD is set to 15.
}
\author{
Guoji Guo, Mikael Huss, Guo Qing Tong, Chaoyang Wang, Li Li Sun, Neil D. Clarke, Paul Robson \email{robsonp@gis.a-star.edu.sg}
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S1534580710001103}
}
\keyword{data}

