% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessor-generics.r
\name{destiny generics}
\alias{dataset}
\alias{dataset<-}
\alias{destiny generics}
\alias{distance}
\alias{distance<-}
\alias{eigenvalues}
\alias{eigenvalues<-}
\alias{eigenvectors}
\alias{eigenvectors<-}
\alias{optimal_sigma}
\alias{sigmas}
\alias{sigmas<-}
\title{destiny generics}
\usage{
eigenvalues(object)

eigenvalues(object) <- value

eigenvectors(object)

eigenvectors(object) <- value

sigmas(object)

sigmas(object) <- value

dataset(object)

dataset(object) <- value

distance(object)

distance(object) <- value

optimal_sigma(object)
}
\value{
\code{eigenvalues} retrieves the numeric eigenvalues

\code{eigenvectors} retrieves the eigenvectors matrix

\code{sigmas} retrieves the \code{\link{Sigmas}} from an object utilizing it as kernel width

\code{dataset} retrieves the data the object was created from

\code{distance} retrieves the distance metric used to create the object, e.g. \code{euclidean}

\code{optimal_sigma} retrieves the numeric value of the optimal sigma or local sigmas
}
\description{
destiny provides several generic methods and implements them for the \code{\link{DiffusionMap}} and \code{\link{Sigmas}} classes.
}
\seealso{
\link{DiffusionMap methods} and \link{Sigmas class} for implementations
}

