% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update-methods.r
\docType{methods}
\name{destiny updateObject methods}
\alias{destiny updateObject methods}
\alias{updateObject,Diffusionmap-method}
\alias{updateObject,sigmas-method}
\title{Update old DiffusionMaps to a newer version}
\usage{
\S4method{updateObject}{DiffusionMap}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A DiffusionMap/Sigmas object}

\item{...}{ignored}

\item{verbose}{tells what is being updated}
}
\description{
Update old DiffusionMaps to a newer version
}

