% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap-methods-accession.r
\docType{methods}
\name{DiffusionMap accessors}
\alias{DiffusionMap accessors}
\alias{dataset,DiffusionMap-method}
\alias{dataset<-,DiffusionMap-method}
\alias{distance,DiffusionMap-method}
\alias{distance<-,DiffusionMap-method}
\alias{eigenvalues,DiffusionMap-method}
\alias{eigenvalues<-,DiffusionMap-method}
\alias{eigenvectors,DiffusionMap-method}
\alias{eigenvectors<-,DiffusionMap-method}
\alias{optimal_sigma,DiffusionMap-method}
\alias{sigmas,DiffusionMap-method}
\alias{sigmas<-,DiffusionMap-method}
\title{DiffusionMap accession methods}
\usage{
\S4method{eigenvalues}{DiffusionMap}(object)

\S4method{eigenvalues}{DiffusionMap}(object) <- value

\S4method{eigenvectors}{DiffusionMap}(object)

\S4method{eigenvectors}{DiffusionMap}(object) <- value

\S4method{sigmas}{DiffusionMap}(object)

\S4method{sigmas}{DiffusionMap}(object) <- value

\S4method{dataset}{DiffusionMap}(object)

\S4method{dataset}{DiffusionMap}(object) <- value

\S4method{distance}{DiffusionMap}(object)

\S4method{distance}{DiffusionMap}(object) <- value

\S4method{optimal_sigma}{DiffusionMap}(object)
}
\arguments{
\item{object}{A DiffusionMap}

\item{value}{Vector of eigenvalues or matrix of eigenvectors to get/set}
}
\value{
The assigned or retrieved value
}
\description{
Get and set eigenvalues, eigenvectors, and sigma(s) of a \link{DiffusionMap} object or print information about a DiffusionMap
}
\examples{
data(guo)
dm <- DiffusionMap(guo)
eigenvalues(dm)
eigenvectors(dm)
sigmas(dm)
dataset(dm)
optimal_sigma(dm)

}
\seealso{
\link{DiffusionMap extraction}, \link{DiffusionMap methods}, \link{coercions} for more methods
}

