% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap.r
\docType{class}
\name{DiffusionMap class}
\alias{DiffusionMap}
\alias{DiffusionMap class}
\alias{DiffusionMap-class}
\title{Create a diffusion map of cells}
\usage{
DiffusionMap(data, sigma = NULL, k = find.dm.k(nrow(data) - 1L),
  n.eigs = min(20L, nrow(data) - 2L), density.norm = TRUE, ...,
  distance = c("euclidean", "cosine", "rankcor"), censor.val = NULL,
  censor.range = NULL, missing.range = NULL, vars = NULL,
  verbose = !is.null(censor.range))
}
\arguments{
\item{data}{Expression data to be analyzed. Provide \code{vars} to select specific columns other than the default: all double value columns}

\item{sigma}{Diffusion scale parameter of the Gaussian kernel. Either a number or a \link{Sigmas} object.
(Optional. default: will be calculated using \link{find.sigmas})
A larger sigma might be necessary if the eigenvalues can not be found because of a singularity in the matrix}

\item{k}{Number of nearest neighbors to consider (default: a guess betweeen 100 and \eqn{n - 1})
\code{NULL} or \code{NA} are also interpreted as \code{n - 1L}.}

\item{n.eigs}{Number of eigenvectors/values to return (default: 20)}

\item{density.norm}{logical. If TRUE, use density normalisation}

\item{...}{All parameter after this are optional and have to be specified by name}

\item{distance}{Distance measurement method. Euclidean distance (default), cosine distance (\eqn{1-corr(c_1, c_2)}) or rank correlation distance (\eqn{1-corr(rank(c_1), rank(c_2))}).}

\item{censor.val}{Value regarded as uncertain. Either a single value or one for every dimension (Optional, default: CENSOR.VAL)}

\item{censor.range}{Uncertainity range for censoring (Optional, default: none). A length-2-vector of certainty range start and end. TODO: also allow \eqn{2\times G} matrix}

\item{missing.range}{Whole data range for missing value model. Has to be specified if NAs are in the data}

\item{vars}{Variables (columns) of the data to use. Specifying NULL will select all columns (default: All floating point value columns)}

\item{verbose}{Show a progressbar and other progress information (default: do it if censoring is enabled)}
}
\value{
A DiffusionMap object:
}
\description{
The provided data can be a double \link[base]{matrix} of expression data or a \link[base]{data.frame} with all non-integer (double) columns
being treated as expression data features (and the others ignored), or an \link[Biobase]{ExpressionSet}.
}
\section{Slots}{

\describe{
\item{\code{eigenvalues}}{Eigenvalues ranking the eigenvectors}

\item{\code{eigenvectors}}{Eigenvectors mapping the datapoints to \code{n.eigs} dimensions}

\item{\code{sigmas}}{\link{Sigmas} object with either information about the \link{find.sigmas} heuristic run or just \link{optimal.sigma}.}

\item{\code{data.env}}{Environment referencing the data used to create the diffusion map}

\item{\code{eigenvec0}}{First (constant) eigenvector not included as diffusion component.}

\item{\code{transitions}}{Transition probabilities}

\item{\code{phi0}}{Density vector of normalized transition probability matrix}

\item{\code{d}}{Density vector of transition probability matrix}

\item{\code{k}}{The k parameter for kNN}

\item{\code{density.norm}}{Was density normalization used?}

\item{\code{distance}}{Distance measurement method used.}

\item{\code{censor.val}}{Censoring value}

\item{\code{censor.range}}{Censoring range}

\item{\code{missing.range}}{Whole data range for missing value model}

\item{\code{vars}}{Vars parameter used to extract the part of the data used for diffusion map creation}
}}
\examples{
data(guo)
DiffusionMap(guo)
DiffusionMap(guo, 13, censor.val = 15, censor.range = c(15, 40), verbose = TRUE)

}
\seealso{
\link{DiffusionMap-methods} to get and set the slots. \code{\link{find.sigmas}} to pre-calculate a fitting \code{sigma} parameter
}

