% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap-methods-extraction.r
\docType{methods}
\name{DiffusionMap extraction}
\alias{$,DiffusionMap-method}
\alias{DiffusionMap extraction}
\alias{[[,DiffusionMap,character,missing-method}
\alias{names,DiffusionMap-method}
\alias{names.DiffusionMap}
\title{DiffusionMap extraction methods}
\usage{
\S4method{names}{DiffusionMap}(x)

\S4method{[[}{DiffusionMap,character,missing}(x, i, j, ...)

\S4method{$}{DiffusionMap}(x, name)
}
\arguments{
\item{x}{\code{\link{DiffusionMap}} object}

\item{i, name}{Name of a diffusion component (DC1,...), or column from the data}

\item{j}{N/A}

\item{...}{ignored}
}
\description{
DiffusionMap extraction methods
}
\seealso{
\link[base]{Extract}, \code{\link[base]{names}} for the generics. \link{DiffusionMap accessors}, \link{DiffusionMap methods}, \link{coercions} for more methods
}

