% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap-methods.r
\docType{methods}
\name{DiffusionMap methods}
\alias{DiffusionMap methods}
\alias{DiffusionMap-methods}
\alias{print,DiffusionMap-method}
\alias{print.DiffusionMap}
\alias{show,DiffusionMap-method}
\alias{show.DiffusionMap}
\title{DiffusionMap methods}
\usage{
\S4method{print}{DiffusionMap}(x)

\S4method{show}{DiffusionMap}(object)
}
\description{
Methods for external operations on diffusion maps
}
\seealso{
\link{DiffusionMap accessors}, \link{DiffusionMap extraction}, \link{DiffusionMap coercions} for more methods
}

