% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt.r
\name{dpt helpers}
\alias{dpt helpers}
\alias{dpt_to_cell}
\title{Calculate DPTs}
\usage{
dpt_to_cell(dm, cell)
}
\arguments{
\item{dm}{A \code{\link{DiffusionMap}} object}

\item{cell}{Index of a cell for which all DPTs should be calculated}
}
\description{
Given a cell index, returns the DPT of this cell to all cells
}

