% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt-methods.r
\docType{methods}
\name{DPT methods}
\alias{DPT methods}
\title{DPT methods}
\usage{
\S4method{dataset}{DPT}(object)

\S4method{dataset}{DPT}(object) <- value
}
\description{
Methods for the \link{DPT} class
}

