% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt.r
\docType{class}
\name{DPT}
\alias{DPT}
\alias{DPT-class}
\title{Diffusion Pseudo Time}
\usage{
DPT(dm, tips = random_root(dm), ..., w_width = 0.1)
}
\arguments{
\item{dm}{A \code{\link{DiffusionMap}} object. Its transition probabilities will be used to calculate the DPT}

\item{tips}{The cell index/indices from which to calculate the DPT(s) (integer of length 1-3)}

\item{w_width}{Window width to use for deciding the branch cutoff}
}
\description{
Create pseudotime ordering and assigns cell to one of three branches
}
\section{Slots}{

\describe{
\item{\code{branch}}{Branch labels for each cell; \code{1:3} or \code{NA} for undeceided}

\item{\code{parent}}{\code{\link{matrix}} of parent branches (may be of \code{ncol(...) == 0})}

\item{\code{dpt}}{Diffusion pseudotime in respect to the root cell (and other tips if \code{branching == TRUE})}

\item{\code{dm}}{\code{\link{DiffusionMap}} used to create this DPT object}
}}
\examples{
data(guo_norm)
dm <- DiffusionMap(guo_norm)
dpt <- DPT(dm, branching = TRUE)
str(dpt)

}

