% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guo-data.r
\docType{data}
\name{guo}
\alias{data:guo}
\alias{data:guo.norm}
\alias{guo}
\alias{guo.norm}
\title{Guo at al. mouse embryonic stem cell qPCR data}
\format{A data frame with 492 rows and 49 variables, the first 48 of which are Ct values for gene expression data}
\usage{
data(guo)
data(guo.norm)
}
\value{
a \eqn{429 \times 49} \code{data.frame} with the first 48 columns being qPCR Ct values and the last one the "divisions" annotation.
}
\description{
Gene expression data of 48 genes and an annotation column \code{$num.cells} containing the cell stage at which the embryos were harvested.
}
\details{
The data is normalized using the mean of two housekeeping genes.
The difference between \code{guo} and \code{guo.norm} is the LoD being set to 10 in the former, making it usable with the \code{censor.val} parameter of \link{DiffusionMap}.
}
\author{
Guoji Guo, Mikael Huss, Guo Qing Tong, Chaoyang Wang, Li Li Sun, Neil D. Clarke, Paul Robson \email{robsonp@gis.a-star.edu.sg}
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S1534580710001103}
}
\keyword{data}

