% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update-methods.r
\docType{methods}
\name{updateObject,DiffusionMap-method}
\alias{updateObject,DiffusionMap-method}
\title{Update old \link{DiffusionMap}s to a newer version}
\usage{
\S4method{updateObject}{DiffusionMap}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A \link{DiffusionMap} object created with an older destiny release}

\item{...}{ignored}

\item{verbose}{tells what is being updated}
}
\value{
A \link{DiffusionMap} object that is valid when used with the current destiny release
}
\description{
Update old \link{DiffusionMap}s to a newer version
}

