% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-update.r
\docType{methods}
\name{updateObject-method}
\alias{updateObject,DiffusionMap-method}
\alias{updateObject,Sigmas-method}
\alias{updateObject-method}
\title{Update old \link{DiffusionMap}s or \link{Sigmas} to a newer version}
\usage{
\S4method{updateObject}{DiffusionMap}(object, ..., verbose = FALSE)

\S4method{updateObject}{Sigmas}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{A \link{DiffusionMap} or \link{Sigmas} object created with an older destiny release}

\item{...}{ignored}

\item{verbose}{tells what is being updated}
}
\value{
A \link{DiffusionMap} or \link{Sigmas} object that is valid when used with the current destiny release
}
\description{
Update old \link{DiffusionMap}s or \link{Sigmas} to a newer version
}

