% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-update.r
\docType{methods}
\name{updateObject methods}
\alias{updateObject methods}
\alias{updateObject.DiffusionMap}
\alias{updateObject.Sigmas}
\alias{updateObject.GeneRelevance}
\alias{updateObject,DiffusionMap-method}
\alias{updateObject,Sigmas-method}
\alias{updateObject,GeneRelevance-method}
\title{Update old destiny objects to a newer version.}
\usage{
\S4method{updateObject}{DiffusionMap}(object, ..., verbose = FALSE)

\S4method{updateObject}{Sigmas}(object, ..., verbose = FALSE)

\S4method{updateObject}{GeneRelevance}(object, ..., verbose = FALSE)
}
\arguments{
\item{object}{An object created with an older destiny release}

\item{...}{ignored}

\item{verbose}{tells what is being updated}
}
\value{
A \link{DiffusionMap} or \link{Sigmas} object that is valid when used with the current destiny release
}
\description{
Handles \link{DiffusionMap}, \link{Sigmas}, and \link[=gene_relevance]{GeneRelevance}.
}
