% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelpers.r
\name{colorlegend}
\alias{colorlegend}
\title{Color legend}
\usage{
colorlegend(col, pal = palette(), log = FALSE, posx = c(0.9, 0.93),
  posy = c(0.05, 0.9), main = NULL, main.cex = 1, main.col = par("fg"),
  lab.col = par("fg"), steps = 5, color.steps = 100, digit = 2,
  left = FALSE, ...)
}
\arguments{
\item{col}{Vector of factor, integer, or double used to determine the ticks.}

\item{pal}{If \code{col} is double, pal is used as a continuous palette, else as categorical one}

\item{log}{Use logarithmic scale?}

\item{posx}{Left and right borders of the color bar relative to plot area (Vector of length 2; 0-1)}

\item{posy}{Bottom and top borders of color bar relative to plot area (Vector of length 2; 0-1)}

\item{main}{Legend title}

\item{main.cex}{Size of legend title font}

\item{main.col}{Color of legend title}

\item{lab.col}{Color of tick or category labels}

\item{steps}{Number of labels in case of a continuous axis}

\item{color.steps}{Number of gradient samples in case of continuous axis}

\item{digit}{Number of digits for continuous axis labels}

\item{left}{logical. If TRUE, invert posx}

\item{...}{Additional parameters for the \link[graphics]{text} call used for labels}
}
\value{
This function is called for the side effect of adding a colorbar to a plot and returns nothing/NULL.
}
\description{
Creates a color legend for a vector used to color a plot. It will use the current \code{\link[grDevices]{palette}()} or the specified \code{pal} as reference.
}
\details{
When passed a factor or integer vector, it will create a discrete legend, whereas a double vector will result in a continuous bar.
}
\examples{
color.data <- 1:6
par(mar = par('mar') + c(0, 0, 0, 3))
plot(sample(6), col = color.data)
colorlegend(color.data)

}

