% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destiny-package.r
\docType{package}
\name{destiny}
\alias{destiny}
\alias{destiny-package}
\title{Create and plot diffusion maps}
\description{
The main function is \code{\link{DiffusionMap}}, which returns an object you can \code{\link{plot}} (\code{\link{plot.DiffusionMap}} is then called).
}
\examples{
demo(destiny, ask = FALSE)

}
