% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap-methods-coercion.r
\docType{methods}
\name{DiffusionMap coercions}
\alias{DiffusionMap coercions}
\alias{as.data.frame,DiffusionMap-method}
\alias{as.data.frame.DiffusionMap}
\alias{fortify.DiffusionMap}
\title{DiffusionMap coercion methods}
\usage{
\S4method{as.data.frame}{DiffusionMap}(x, row.names = NULL,
  optional = FALSE, ...)

fortify.DiffusionMap(model, data, ...)
}
\arguments{
\item{x, model}{A \code{\link{DiffusionMap}}}

\item{row.names}{NULL or a character vector giving the row names for the data frame. Missing values are not allowed.}

\item{optional}{logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional.}

\item{...}{Other parameters}

\item{data}{ignored}
}
\value{
An object of the desired class
}
\description{
Convert a \code{\link{DiffusionMap}} to other classes
}
\details{
\link[ggplot2]{fortify} is a ggplot2 generic allowing a diffusion map to be used as \code{data} parameter in \link[ggplot2]{ggplot} or \link[ggplot2]{qplot}.
}
\examples{
data(guo)
dm <- DiffusionMap(guo)
dm$DC1   # A diffusion component
dm$Actb  # A gene expression

}
\seealso{
\link{DiffusionMap accessors}, \link{DiffusionMap extraction}, \link{DiffusionMap methods} for more methods
}

