% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diffusionmap-methods.r
\docType{methods}
\name{DiffusionMap methods}
\alias{DiffusionMap methods}
\alias{DiffusionMap-methods}
\alias{dataset}
\alias{dataset<-}
\alias{eigenvalues}
\alias{eigenvalues<-}
\alias{eigenvectors}
\alias{eigenvectors<-}
\alias{optimal.sigma,DiffusionMap-method}
\alias{print,DiffusionMap-method}
\alias{show,DiffusionMap-method}
\alias{sigmas}
\alias{sigmas<-}
\alias{values}
\alias{values<-}
\alias{vectors}
\alias{vectors<-}
\title{DiffusionMap methods}
\usage{
eigenvalues(dm)

eigenvalues(dm) <- value

eigenvectors(dm)

eigenvectors(dm) <- value

sigmas(dm)

sigmas(dm) <- value

dataset(dm)

dataset(dm) <- value

\S4method{optimal.sigma}{DiffusionMap}(object)

\S4method{print}{DiffusionMap}(x)

\S4method{show}{DiffusionMap}(object)
}
\arguments{
\item{dm,x,object}{A DiffusionMap}

\item{value}{Vector of eigenvalues or matrix of eigenvectors to get/set}
}
\value{
The assigned or retrieved value
}
\description{
Get and set eigenvalues, eigenvectors, and sigma(s) of a \link{DiffusionMap} object or print information about a DiffusionMap
}
\examples{
data(guo)
dm <- DiffusionMap(guo)
eigenvalues(dm)
eigenvectors(dm)
sigmas(dm)
dataset(dm)
optimal.sigma(dm)
}
\seealso{
\code{\link{DiffusionMap}}
}

