% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusionmap-methods.r
\docType{methods}
\name{DiffusionMap methods}
\alias{DiffusionMap methods}
\alias{DiffusionMap-methods}
\alias{print,DiffusionMap-method}
\alias{print.DiffusionMap}
\alias{show,DiffusionMap-method}
\alias{show.DiffusionMap}
\title{DiffusionMap methods}
\usage{
\S4method{print}{DiffusionMap}(x)

\S4method{show}{DiffusionMap}(object)
}
\arguments{
\item{x, object}{A \code{\link{DiffusionMap}}}
}
\value{
The \code{DiffusionMap} object (\code{print}), or NULL (\code{show}), invisibly
}
\description{
Methods for external operations on diffusion maps
}
\examples{
data(guo)
dm <- DiffusionMap(guo)
print(dm)
show(dm)

}
\seealso{
\link{DiffusionMap accessors}, \link{extractions}, \link{coercions} for more methods
}

