% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt-methods-matrix.r
\docType{methods}
\name{DPT matrix methods}
\alias{DPT matrix methods}
\alias{[.DPT}
\alias{dim.DPT}
\alias{ncol.DPT}
\alias{nrow.DPT}
\title{DPT Matrix methods}
\usage{
\S4method{[}{DPT,index,index,logicalOrMissing}(x, i, j, ..., drop = TRUE)

\S4method{[}{DPT,index,missing,logicalOrMissing}(x, i, j, ..., drop = TRUE)

\S4method{[}{DPT,missing,index,logicalOrMissing}(x, i, j, ..., drop = TRUE)

\S4method{[}{DPT,missing,missing,logicalOrMissing}(x, i, j, ..., drop = TRUE)

\S4method{nrow}{DPT}(x)

\S4method{ncol}{DPT}(x)

\S4method{dim}{DPT}(x)
}
\description{
Treat DPT object as a matrix of cell-by-cell DPT distances.
}

