% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt.r
\docType{class}
\name{DPT}
\alias{DPT}
\alias{DPT-class}
\title{Diffusion Pseudo Time}
\usage{
DPT(dm, branching = length(tips) > 1L, tips = if (branching) find_tips(dm,
  root) else root, root = random_root(dm))
}
\arguments{
\item{dm}{A \code{\link{DiffusionMap}} object}

\item{branching}{Detect a branching? (\code{TRUE} or \code{FALSE})}

\item{tips}{Tip cell indices for each branch (integer vector of length 1 or 3)}

\item{root}{The root index from which to calculate the DPTs (integer of length 1)}
}
\value{
A \code{\link[base]{data.frame}} with the rows: \describe{
	\item{\code{Branch}}{Branch labels for each cell, 1,2,3 or NA for undeceided}
 \item{\code{DPT}}{Diffusion pseudotime in respect to the root cell}
 \item{(\code{if (branching)}) \code{DPT.1}, \code{DPT.2}}{Diffusion pseudotime in respect to the other tips}
}
}
\description{
Create pseudotime ordering and assigns cell to one of three branches
}
\details{
All parameters are optional, but at least \code{branching} or \code{tips} has to be there.
If unspecified: \describe{
 \item{\code{root}}{will be the furthest cell from a random one}
	\item{\code{branching}}{will be \code{TRUE} if multiple \code{tips} are specified}
 \item{\code{tips}}{will be \code{root} and (\code{if (branching)}) the most distant cells from it}
}
}

