% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expressionset-helpers.r
\docType{methods}
\name{ExpressionSet helpers}
\alias{ExpressionSet helpers}
\alias{as.ExpressionSet}
\alias{as.ExpressionSet,data.frame-method}
\alias{as.ExpressionSet-method}
\alias{read.ExpressionSet}
\title{Convert object to \link[Biobase]{ExpressionSet} or read it from a file}
\usage{
as.ExpressionSet(x, ...)

\S4method{as.ExpressionSet}{data.frame}(x, annotation.cols = !sapply(x,
  is.double))

read.ExpressionSet(file, header = TRUE, ...)
}
\arguments{
\item{x}{\link[base]{data.frame} to convert to an \link[Biobase]{ExpressionSet}.}

\item{...}{Additional parameters to \link[utils]{read.table}}

\item{annotation.cols}{The data.frame columns used as annotations. All others are used as expressions. (Logical, character or numerical index array)}

\item{file}{File path to read ASCII data from}

\item{header}{Specifies if the file has a header row.}
}
\value{
an \link[Biobase]{ExpressionSet} object
}
\description{
These functions present quick waya to create \link[Biobase]{ExpressionSet} objects.
}
\details{
They work by using all continuous (double) columns as expression data, and all others as sample annotations.
}
\examples{
df <- data.frame(Time  = seq_len(3), #integer column
                 Actb  = c(0.05, 0.3, 0.8),
                 Gapdh = c(0.2, 0.03, 0.1))
set <- as.ExpressionSet(df)
rownames(exprs(set)) == c('Actb', 'Gapdh')
phenoData(set)$Time == 1:3
}
\seealso{
\link[utils]{read.table} on which \code{read.ExpressionSet} is based, and \link[Biobase]{ExpressionSet}.
}

