% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-relevance-plotting.r
\docType{methods}
\name{Gene Relevance plotting}
\alias{Gene Relevance plotting}
\alias{plot.GeneRelevance}
\alias{plot_gradient_map}
\alias{plot_gene_relevance}
\title{Plot gene relevance or gradient map}
\usage{
\S4method{plot}{GeneRelevance,character}(x, y, ...)

\S4method{plot}{GeneRelevance,numeric}(x, y, ...)

\S4method{plot}{GeneRelevance,missing}(x, y, ...)

plot_gradient_map(dm, gene, ..., pal = cube_helix)

plot_gene_relevance(dm, iter_smooth = 2L, genes = 5L, ...,
  pal = palette())
}
\arguments{
\item{x}{\code{\link{GeneRelevance}} object.}

\item{y, gene}{Gene name or index to create gradient map for. (a number or string)}

\item{...}{Passed to \code{plot_gradient_map}/\code{plot_gene_relevance}}

\item{dm}{\code{\link{DiffusionMap}} object.}

\item{pal}{Palette. Either A colormap function or a list of colors}

\item{iter_smooth}{Number of label smoothing iterations to perform on relevance map.
The higher the more homogenous and the less local structure.}

\item{genes}{Genes to based relevance map on or number of genes to use. (vector of strings or one number)
You can also pass an index into the gene names. (vector of numbers or logicals with length > 1)}
}
\value{
ggplot2 plot, when plotting a relevance map with a list member $ids containing the IDs used.
}
\description{
\code{plot(gene_relevance, 'Gene')} plots the gradient map of this gene,
\code{plot(gene_relevance)} a relevance map of a selection of genes.
}
