% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-relevance.r
\docType{class}
\name{Gene Relevance}
\alias{Gene Relevance}
\alias{GeneRelevance}
\alias{GeneRelevance-class}
\alias{gene_relevance}
\alias{gene_relevance,matrix,matrix-method}
\alias{gene_relevance,DiffusionMap,missing-method}
\title{Gene relevances for entire data set}
\usage{
gene_relevance(coords, exprs, ..., k = 20L, dims = 1:2, distance = NULL,
  smooth = TRUE, verbose = FALSE)

\S4method{gene_relevance}{DiffusionMap,missing}(coords, exprs, ..., k = 20L,
  dims = 1:2, distance = NULL, smooth = TRUE, verbose = FALSE)

\S4method{gene_relevance}{matrix,matrix}(coords, exprs, ..., k = 20L,
  dims = 1:2, distance = NULL, smooth = TRUE, verbose = FALSE)
}
\arguments{
\item{coords}{A \code{\link{DiffusionMap}} object or a cells \eqn{\times} dims \code{\link{matrix}}.}

\item{exprs}{An cells \eqn{\times} genes \code{\link{matrix}}. Only provide if \code{coords} is no \code{\link{DiffusionMap}}.}

\item{...}{If no \code{\link{DiffusionMap}} is provided, a vector of \code{weights} (of the same length as \code{dims}) can be provided.}

\item{k}{Number of nearest neighbors to use}

\item{dims}{Index into columns of \code{coord}}

\item{distance}{Distance measure to use for the nearest neighbor search.}

\item{smooth}{Smoothing parameters \code{c(window, alpha)} (see \code{\link[smoother]{smth.gaussian}}).
Alternatively \code{\link{TRUE}} to use the \link{smoother} \link[smoother:smth.options]{defaults} or \code{\link{FALSE}} to skip smoothing,}

\item{verbose}{If TRUE, log additional info to the console}
}
\value{
A \code{GeneRelevance} object:
}
\description{
The relevance map is cached insided of the \code{\link{DiffusionMap}}.
}
\section{Slots}{

\describe{
\item{\code{coords}}{A cells \eqn{\times} dims \code{\link{matrix}} of coordinates (e.g. diffusion components), reduced to the dimensions passed as \code{dims}}

\item{\code{exprs}}{A cells \eqn{\times} genes matrix of expressions}

\item{\code{partials}}{Array of partial derivatives wrt to considered dimensions in reduced space
(genes \eqn{\times} cells \eqn{\times} dimensions)}

\item{\code{partials_norm}}{Matrix with norm of aforementioned derivatives. (n\_genes \eqn{\times} cells)}

\item{\code{nn_index}}{Matrix of k nearest neighbor indices. (cells \eqn{\times} k)}

\item{\code{dims}}{Column index for plotted dimensions. Can \code{\link{character}}, \code{\link{numeric}} or \code{\link{logical}}}

\item{\code{distance}}{Distance measure used in the nearest neighbor search. See \code{\link{find_knn}}}

\item{\code{smooth_window}}{Smoothing window used (see \code{\link[smoother]{smth.gaussian}})}

\item{\code{smooth_alpha}}{Smoothing kernel width used (see \code{\link[smoother]{smth.gaussian}})}
}}

\examples{
data(guo_norm)
dm <- DiffusionMap(guo_norm)
gr <- gene_relevance(dm)

m <- t(Biobase::exprs(guo_norm))
gr_pca <- gene_relevance(prcomp(m)$x, m)
# now plot them!

}
\seealso{
\link{Gene Relevance methods}, \link{Gene Relevance plotting}: \code{plot_differential_map}/\code{plot_gene_relevance}
}
