% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene-relevance.r
\docType{class}
\name{Gene Relevance}
\alias{gene_relevance}
\title{Gene relevances for entire data set}
\usage{
gene_relevance(dm, k = 20, dims = 1:2, n_proc = 1)
}
\arguments{
\item{dm}{A \code{\link{DiffusionMap}} object.}

\item{dims}{Index into columns of \code{coord}}

\item{n_proc}{Number of threads for parallelisation.}
}
\value{
A \code{GeneRelevance} object:
}
\description{
The relevance map is cached insided of the \code{\link{DiffusionMap}}.
}
\section{Slots}{

\describe{
\item{\code{partial}}{Array of partial derivatives wrt to considered dimensions in reduced space.
(genes \eqn{\times} cells \eqn{\times} dimensions)}

\item{\code{partials_norm}}{Matrix with norm of aforementioned derivatives. (n\_genes \eqn{\times} cells)}

\item{\code{nn_index}}{Matrix of k nearest neighbor indices. (cells \eqn{\times} k)}
}}

\seealso{
\link{Gene Relevance plotting}: \code{plot_gradient_map}/\code{plot_gene_relevance}
}
