% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guo-data.r
\docType{data}
\name{guo}
\alias{guo}
\alias{data:guo}
\alias{data:guo_norm}
\alias{guo_norm}
\title{Guo at al. mouse embryonic stem cell qPCR data}
\format{An \link[Biobase]{ExpressionSet} with 48 features, 428 samples and 2 \link[Biobase]{phenoData} annotations.}
\usage{
data(guo)
data(guo_norm)
}
\value{
an \link[Biobase]{ExpressionSet} with 48 features and 428 samples containing qPCR Ct values and a "num.cells" sample annotation.
}
\description{
Gene expression data of 48 genes and an annotation column \code{$num_cells} containing the cell stage at which the embryos were harvested.
}
\details{
The data is normalized using the mean of two housekeeping genes.
The difference between \code{guo} and \code{guo_norm} is the LoD being set to 10 in the former, making it usable with the \code{censor_val} parameter of \link{DiffusionMap}.
}
\references{
\url{http://www.sciencedirect.com/science/article/pii/S1534580710001103}
}
\author{
Guoji Guo, Mikael Huss, Guo Qing Tong, Chaoyang Wang, Li Li Sun, Neil D. Clarke, Paul Robson \email{robsonp@gis.a-star.edu.sg}
}
\keyword{data}
