% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt-plotting.r
\docType{methods}
\name{plot.DPT}
\alias{plot,DPT,missing-method}
\alias{plot,DPT,numeric-method}
\alias{plot.DPT}
\title{Plot DPT}
\usage{
plot.DPT(x, root = 1L, paths_to = integer(0L), dcs = 1:2,
  divide = integer(0L), w_width = 0.1, col_by = "dpt",
  col_path = palette(), col_tip = "red", ..., pal = if
  (is.double(x[[col_by]])) cube_helix else palette(), draw_legend = TRUE)

\S4method{plot}{DPT,numeric}(x, y, ...)

\S4method{plot}{DPT,missing}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{DPT}} object.}

\item{paths_to}{Numeric Branch IDs. Are used as target(s) for the path(s) to draw.}

\item{divide}{If \code{col_by = 'branch'}, this specifies which branches to divide. (see \code{\link{branch_divide}})}

\item{w_width}{Window width for smoothing the path (see \code{\link[smoother]{smth.gaussian}})}

\item{col_by}{Color by 'dpt' (DPT starting at \code{branches[[1]]}), 'branch', or a veriable of the data.}

\item{col_path}{Colors for the path or a function creating n colors}

\item{col_tip}{Color for branch tips}

\item{...}{Graphical parameters supplied to \code{\link{plot.DiffusionMap}}}

\item{pal}{Palette to use for coloring the points}

\item{y, root}{Root branch ID. Will be used as the start of the DPT.
(If longer than size 1, will be interpreted as \code{c(root, branches)})}

\item{dcx, dcy}{The dimensions to use from the DiffusionMap}
}
\description{
Plots diffusion components from a Diffusion Map and the accompanying Diffusion Pseudo Time (\code{\link{DPT}})
}
\examples{
data(guo_norm)
dm <- DiffusionMap(guo_norm)
dpt <- DPT(dm, branching = TRUE)
plot(dpt)
plot(dpt, 2L,      col_by = 'branch')
plot(dpt, 1L, 2:3, col_by = 'num_cells')

}

