% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt_plotting.r
\docType{methods}
\name{plot.DPT}
\alias{plot,DPT,missing-method}
\alias{plot,DPT,numeric-method}
\alias{plot.DPT}
\title{Plot DPT}
\usage{
plot.DPT(x, branches = 1L, dcx = 1L, dcy = 2L, w_width = 0.1,
  col_by = "dpt", col_map = cube_helix, col_path = palette(),
  col_tip = "red", ...)

\S4method{plot}{DPT,numeric}(x, y, ...)

\S4method{plot}{DPT,missing}(x, y, ...)
}
\arguments{
\item{x}{A \code{\link{DPT}} object.}

\item{dcx, dcy}{The dimensions to use from the DiffusionMap}

\item{w_width}{Window width for smoothing the path (see \code{\link[smoother]{smth.gaussian}})}

\item{col_by}{Color by 'dpt' (DPT starting at \code{branches[[1]]}), 'branch', or a veriable of the data.}

\item{col_map}{Color Map for DPT coloring. Sequence of colors or a function accepting the number of colors to create}

\item{col_path}{Colors for the path or a function creating n colors}

\item{col_tip}{Color for branch tips}

\item{...}{Graphical parameters supplied to \code{\link{plot.DiffusionMap}}}

\item{y, branches}{Numeric Branch IDs. The first one will be used as the start of the DPT, subsequent ones for the path(s).}
}
\description{
Plots diffusion components from a Diffusion Map and the accompanying Diffusion Pseudo Time (\code{\link{DPT}})
}
\examples{
data(guo_norm)
dm <- DiffusionMap(guo_norm)
dpt <- DPT(dm, branching = TRUE)
plot(dpt)
plot(dpt, 2L,  col_by = 'branch')
plot(dpt, 1:3, col_by = 'num_cells')

}

