% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection-dist.r
\name{projection_dist}
\alias{projection_dist}
\title{Projection distance}
\usage{
projection_dist(dists_or_data, new_data, ...)
}
\arguments{
\item{dists_or_data}{If new_data is not preset, this needs to be a \code{\link[=proxy::dist]{crossdist}} object}

\item{new_data}{A \code{\link[base]{matrix}} \code{\link[Matrix]{Matrix}}, or \code{\link[base]{data.frame}}.
If given, \code{dists_or_data e} needs to be of one of the same classes.}

\item{...}{Passed to \code{\link{proxa::dist}}}
}
\description{
Projection distance
}

