% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpt_helpers.r
\name{random_root}
\alias{random_root}
\title{Find a random root cell index}
\usage{
random_root(dm)
}
\arguments{
\item{dm}{A \code{\link{DiffusionMap}} object}
}
\description{
Finds a cell index whose cell has the maximum DPT distance from a randomly selected one.
}

